Move Pagination and TimeRange to the kernel module
