+++ delta/kernel/src/main/scala/ch/epfl/bluebrain/nexus/delta/kernel/search/Pagination.scala
@@ -1,4 +1,4 @@
-package ch.epfl.bluebrain.nexus.delta.sdk.model.search
+package ch.epfl.bluebrain.nexus.delta.kernel.search
 
 import io.circe.Json
